set the view as listenner for controller (to finish at home...)
