+++ ClientApp/src/app/controllers/WelcomeController.java
@@ -3,6 +3,11 @@ package app.controllers;
 import app.models.User;
 import app.helpers.Serialization;
 
+import java.io.IOException;
+import java.net.Socket;
+import java.net.UnknownHostException;
+import java.util.ResourceBundle;
+
 import java.awt.Image;
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -11,16 +16,43 @@ import java.io.PrintWriter;
 import java.net.Socket;
 
 public class WelcomeController {
+  private WelcomeListener listener;
+
+  private String url;
+  private int port;
   private static Socket socket;
   private PrintWriter out = null;
   private BufferedReader in = null;
   private String serverAnswer;
 
-  public WelcomeController(Socket sck) {
-    this.socket = sck;
+  //answer.setAlignmentX(Component.CENTER_ALIGNMENT);
+
+  public WelcomeController(String url, int port) {
+    this.url = url;
+    this.port = port;
   }
 
-  public String connection(String login, String pwd ) {
+  public void addListener(WelcomeListener l) {
+    listener = l;
+  }
+
+  private void createSocket(){
+    String answer;
+    try {
+      socket = new Socket(url, port);
+      listener.authenticationIhm();
+    } catch (Exception e) {
+      if (e instanceof UnknownHostException)
+        answer = "Impossible de se connecter à l'adresse " + socket.getLocalAddress();
+      else
+        answer = "Aucun serveur à l'écoute du port";
+
+      l.updateAnswerLabel(answer);
+    }
+  }
+
+  public void getConnection(String login, String pwd ) {
+    Strign serverAnswer;
     try {
       out = new PrintWriter(socket.getOutputStream());
       in = new BufferedReader(new InputStreamReader(socket.getInputStream()));
@@ -33,12 +65,11 @@ public class WelcomeController {
 
       //Waiting for the answer (answer = "authentic" if success)
       serverAnswer=in.readLine();
-      if (serverAnswer.equals("authentic")){
-
-      }
     } catch (Exception e) {
       serverAnswer = "disconnected";
     }
-    return serverAnswer;
+
+    for (HelloListener l : listeners)
+      l.TATATAA(serverAnswer);
   }
 }
