+++ ClientApp/src/app/listeners/WelcomeListener.java
@@ -1,0 +1,6 @@
+import java.util.*;
+
+public interface WelcomeListener {
+    void authenticationIhm();
+    void updateAnswerLabel(String answer);
+}
